Send MDPW_DISCONNECT when worker is destroyed
As discussed in #45, the worker does not currently provide a mechanism
for informing the broker that it is going away, which may result in
messages routed to the destroyed broker in the time between when it
is destroyed and its heartbeat expires on the broker.  To resolve this,
send MDPW_DISCONNECT when the worker is destroyed.

Note:  This implementation unconditionally sends MDPW_DISCONNECT in
mdp_worker_destroy rather than adding an mdp_worker_destroy function
because MDPW_DISCONNECT should be harmless in all states in which it is
not expected, the overhead should be small, and the adding a
disconnected-but-not-destroyed worker state seemed to add more
complexity than its value.

Fixes #45.

Signed-off-by: Kevin Locke <klocke@quantpost.com>