Add a default linger time and expose it
In order for the DISCONNECT message to be sent, the socket needs to send
data before being destroyed.  The czmq default linger time of 0 won't
work for this case.  Bump it up to 100ms, which should be sufficient, if
there isn't another large send operation queued, but shouldn't
significantly alter the behavior if the broker has gone away and the
messages can't be sent.

Note:  Personally, I contend that a default linger of -1 is closer to
the expected behavior for this sort of protocol, but my assessment was
that it would introduce enough backwards-compatibility issues that a
small value was more reasonable.  Feel free to change this in future
commits if it is deemed less of an issue.

Signed-off-by: Kevin Locke <klocke@quantpost.com>